Rename editor-toolbar to components
