+++ ts/components/ColorPicker.d.ts
