+++ ts/components/LabelButton.svelte
