+++ ts/components/SelectOption.svelte
