+++ ts/components/SquareButton.svelte
