+++ ts/components/StickyBar.svelte
