+++ ts/components/WithShortcut.svelte
