+++ ts/components/hideable.ts
