+++ ts/editor/BUILD.bazel
@@ -1,9 +1,27 @@
 load("@npm//@bazel/typescript:index.bzl", "ts_library")
+load("//ts/svelte:svelte.bzl", "compile_svelte", "svelte_check")
 load("//ts:prettier.bzl", "prettier_test")
 load("//ts:eslint.bzl", "eslint_test")
 load("//ts:esbuild.bzl", "esbuild")
 load("//ts:vendor.bzl", "copy_bootstrap_icons", "copy_mdi_icons")
 load("//ts:compile_sass.bzl", "compile_sass")
+
+svelte_files = glob(["*.svelte"])
+
+svelte_names = [f.replace(".svelte", "") for f in svelte_files]
+
+filegroup(
+    name = "svelte_components",
+    srcs = svelte_names,
+    visibility = ["//visibility:public"],
+)
+
+compile_svelte(
+    name = "svelte",
+    srcs = svelte_files,
+    deps = [],
+    visibility = ["//visibility:public"],
+)
 
 compile_sass(
     srcs = [
@@ -30,6 +48,19 @@ compile_sass(
     ],
 )
 
+compile_sass(
+    srcs = [
+        "bootstrap.scss",
+        "legacy.scss",
+    ],
+    group = "local_css",
+    deps = [
+        "//ts/sass:button_mixins_lib",
+        "//ts/sass/bootstrap",
+    ],
+    visibility = ["//visibility:public"],
+)
+
 ts_library(
     name = "editor_ts",
     srcs = glob(["*.ts"]),
@@ -38,8 +69,8 @@ ts_library(
         "//ts:image_module_support",
         "//ts/lib",
         "//ts/sveltelib",
+        "//ts/components",
         "//ts/html-filter",
-        "//ts/editor-toolbar",
         "@npm//svelte",
     ],
 )
@@ -106,8 +137,9 @@ esbuild(
         "bootstrap-icons",
         "mdi-icons",
         "editor_ts",
-        "//ts/editor-toolbar:local_css",
-        "//ts/editor-toolbar:svelte_components",
+        "local_css",
+        "svelte_components",
+        "//ts/components:svelte_components",
     ],
 )
 
@@ -129,3 +161,15 @@ prettier_test(
 #         ],
 #     ),
 # )
+
+svelte_check(
+    name = "svelte_check",
+    srcs = glob([
+        "*.ts",
+        "*.svelte",
+    ]) + [
+        "//ts/sass:button_mixins_lib",
+        "//ts/sass/bootstrap",
+        "@npm//@types/bootstrap",
+    ],
+)
